Moved declaration of individual proeprties to ceparate classes by family
